+++ GalaxyServer/XnaGeometry/MathHelper.cs
@@ -33,32 +33,32 @@ namespace XnaGeometry
 {
     public static class MathHelper
     {
-        public const double E = (double)Math.E;
-        public const double Log10E = 0.4342945f;
-        public const double Log2E = 1.442695f;
-        public const double Pi = (double)Math.PI;
-        public const double PiOver2 = (double)(Math.PI / 2.0);
-        public const double PiOver4 = (double)(Math.PI / 4.0);
-        public const double TwoPi = (double)(Math.PI * 2.0);
+        public const float E = (float)Math.E;
+        public const float Log10E = 0.4342945f;
+        public const float Log2E = 1.442695f;
+        public const float Pi = (float)Math.PI;
+        public const float PiOver2 = (float)(Math.PI / 2.0);
+        public const float PiOver4 = (float)(Math.PI / 4.0);
+        public const float TwoPi = (float)(Math.PI * 2.0);
         
-        public static double Barycentric(double value1, double value2, double value3, double amount1, double amount2)
+        public static float Barycentric(float value1, float value2, float value3, float amount1, float amount2)
         {
             return value1 + (value2 - value1) * amount1 + (value3 - value1) * amount2;
         }
 
-        public static double CatmullRom(double value1, double value2, double value3, double value4, double amount)
+        public static float CatmullRom(float value1, float value2, float value3, float value4, float amount)
         {
             // Using formula from http://www.mvps.org/directx/articles/catmull/
-            // Internally using doubles not to lose precission
-            double amountSquared = amount * amount;
-            double amountCubed = amountSquared * amount;
-            return (double)(0.5 * (2.0 * value2 +
+            // Internally using floats not to lose precission
+            float amountSquared = amount * amount;
+            float amountCubed = amountSquared * amount;
+            return (float)(0.5 * (2.0 * value2 +
                 (value3 - value1) * amount +
                 (2.0 * value1 - 5.0 * value2 + 4.0 * value3 - value4) * amountSquared +
                 (3.0 * value2 - value1 - 3.0 * value3 + value4) * amountCubed));
         }
 
-        public static double Clamp(double value, double min, double max)
+        public static float Clamp(float value, float min, float max)
         {
             // First we check to see if we're greater than the max
             value = (value > max) ? max : value;
@@ -70,18 +70,18 @@ namespace XnaGeometry
             return value;
         }
         
-        public static double Distance(double value1, double value2)
+        public static float Distance(float value1, float value2)
         {
             return Math.Abs(value1 - value2);
         }
         
-        public static double Hermite(double value1, double tangent1, double value2, double tangent2, double amount)
+        public static float Hermite(float value1, float tangent1, float value2, float tangent2, float amount)
         {
-            // All transformed to double not to lose precission
+            // All transformed to float not to lose precission
             // Otherwise, for high numbers of param:amount the result is NaN instead of Infinity
-            double v1 = value1, v2 = value2, t1 = tangent1, t2 = tangent2, s = amount, result;
-            double sCubed = s * s * s;
-            double sSquared = s * s;
+            float v1 = value1, v2 = value2, t1 = tangent1, t2 = tangent2, s = amount, result;
+            float sCubed = s * s * s;
+            float sSquared = s * s;
 
             if (amount == 0f)
                 result = value1;
@@ -92,59 +92,59 @@ namespace XnaGeometry
                     (3 * v2 - 3 * v1 - 2 * t1 - t2) * sSquared +
                     t1 * s +
                     v1;
-            return (double)result;
+            return (float)result;
         }
         
         
-        public static double Lerp(double value1, double value2, double amount)
+        public static float Lerp(float value1, float value2, float amount)
         {
             return value1 + (value2 - value1) * amount;
         }
 
-        public static double Max(double value1, double value2)
+        public static float Max(float value1, float value2)
         {
             return Math.Max(value1, value2);
         }
         
-        public static double Min(double value1, double value2)
+        public static float Min(float value1, float value2)
         {
             return Math.Min(value1, value2);
         }
         
-        public static double SmoothStep(double value1, double value2, double amount)
+        public static float SmoothStep(float value1, float value2, float amount)
         {
             // It is expected that 0 < amount < 1
             // If amount < 0, return value1
             // If amount > 1, return value2
 #if(USE_FARSEER)
-            double result = SilverSpriteMathHelper.Clamp(amount, 0f, 1f);
+            float result = SilverSpriteMathHelper.Clamp(amount, 0f, 1f);
             result = SilverSpriteMathHelper.Hermite(value1, 0f, value2, 0f, result);
 #else
-            double result = MathHelper.Clamp(amount, 0f, 1f);
+            float result = MathHelper.Clamp(amount, 0f, 1f);
             result = MathHelper.Hermite(value1, 0f, value2, 0f, result);
 #endif
             return result;
         }
         
-        public static double ToDegrees(double radians)
+        public static float ToDegrees(float radians)
         {
-            // This method uses double precission internally,
-            // though it returns single double
+            // This method uses float precission internally,
+            // though it returns single float
             // Factor = 180 / pi
-            return (double)(radians * 57.295779513082320876798154814105);
+            return (float)(radians * 57.295779513082320876798154814105);
         }
         
-        public static double ToRadians(double degrees)
+        public static float ToRadians(float degrees)
         {
-            // This method uses double precission internally,
-            // though it returns single double
+            // This method uses float precission internally,
+            // though it returns single float
             // Factor = pi / 180
-            return (double)(degrees * 0.017453292519943295769236907684886);
+            return (float)(degrees * 0.017453292519943295769236907684886);
         }
 
-	public static double WrapAngle(double angle)
+	public static float WrapAngle(float angle)
 	{
-		angle = (double)Math.IEEERemainder((double)angle, 6.2831854820251465);
+		angle = (float)Math.IEEERemainder((float)angle, 6.2831854820251465);
 		if (angle <= -3.14159274f)
 		{
 			angle += 6.28318548f;
